Rename datasets/fit.py to datasets/map.py
