+++ gammapy/datasets/__init__.py
@@ -1,3 +1,3 @@
 from .core import *
 from .spectrum import *
-from .fit import *
+from .map import *
