+++ gammapy/tests/test_map.py
